+++ blakserv/commcli.h
@@ -27,9 +27,9 @@ void AddByteToPacket(unsigned char byte1);
 void AddShortToPacket(short byte2);
 void AddIntToPacket(int byte4);
 void AddStringToPacket(int int_len,const char *ptr);
-void SecurePacketBufferList(int session_id,buffer_node *blist);
-void SendPacket(int session_id);
-void SendCopyPacket(int session_id);
+void SecurePacketBufferList(__int64 session_id,buffer_node *blist);
+void SendPacket(__int64 session_id);
+void SendCopyPacket(__int64 session_id);
 void ClearPacket(void);
 
 void ClientHangupToBlakod(session_node *session);
